requirements.txt
# Core RAG and AI Libraries
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Document Processing
PyMuPDF>=1.23.0
python-docx>=0.8.11
python-dotenv>=1.0.0

# Embeddings and Vector Database (Simplified for compatibility)
numpy>=1.24.0

# LLM APIs
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0

# Text Processing
tiktoken>=0.5.0
nltk>=3.8.1

# Web Interface
streamlit>=1.28.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Optional: For better performance
torch>=2.0.0
transformers>=4.30.0
tf-keras>=2.15.0
